Count sums using the parallel addition trick
Tested on i5-6600k. This is the fastest without optimization and
pretty close on -O2 and -O3. For some reason, -O2 is slightly faster
than -O3.

The amount of bitwise operation is horrible and isn't really optimized
at assembly level, probably some kind of superscalarity happens in the
CPU.

This could be easily converted to AVX, but I consider it as
semi-assembly and against the rules.